Week_6/requirements.txt
Flask==3.0.0
azure-cognitiveservices-speech==1.34.0
python-dotenv==1.0.0